Initial Heading refactor
