+++ components/Heading/Heading.css
@@ -30,6 +30,51 @@
   color: var(--white);
 }
 
+.icon {
+  width: 18px;
+  pointer-events: none;
+}
+
+.iconFillWhite {
+  fill: #f7f7f7;
+}
+
+.iconFillBlue {
+  fill: #252e3e;
+}
+
+.iconVisible {
+  visibility: visible;
+}
+
+.iconHidden {
+  visibility: hidden;
+}
+
+:target {
+  content: "";
+  display: block;
+  height: 100px;
+  margin-top: -100px;
+  visibility: hidden;
+}
+
+.anchorDefault {
+  float: left;
+  line-height: 2.25;
+  height: 36px;
+  margin-right: 10px;
+  outline: none;
+}
+
+.anchorOffsetLineHeightOne {
+  float: left;
+  line-height: 1;
+  height: 36px;
+  margin-right: 10px;
+  outline: none;
+}
+
 @media screen and (--medium-viewport) {
   .Heading {
     width: 90%;
